requirements.txt
Django==4.2.9
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
Pillow==10.1.0
python-decouple==3.8
asgiref==3.7.2
daphne==4.0.0
